CEMSChoice {psychotree}	R Documentation

CEMS University Choice Data

Description

Preferences of 303 students from WU Wien for different CEMS universities.

Usage

data("CEMSChoice")

Format

A data frame containing 303 observations on 10 variables.

preference

Paired comparison of class paircomp. Preferences for all 15 paired comparisons from 6 objects: London, Paris, Milano, St. Gallen, Barcelona, Stockholm.

study

Factor coding main discipline of study: commerce, or other (economics, business administration, business education).

english

Factor coding knowledge of English (good, poor).

french

Factor coding knowledge of French (good, poor).

spanish

Factor coding knowledge of Spanish (good, poor).

italian

Factor coding knowledge of Italian (good, poor).

work

Factor. Was the student working full-time while studying?

gender

Factor coding gender.

intdegree

Factor. Does the student intend to take an international degree?

preference1998

Paired comparison of class paircomp. This is like preference but the comparisons between Barcelona an Stockholm are (erroneously) reversed, see below.

Details

Students at Wirtschaftsuniversität Wien (https://www.wu.ac.at/) can study abroad visiting one of currently 17 CEMS universities (Community of European Management Schools and International Companies). Dittrich et al. (1998) conduct and analyze a survey of 303 students to examine the student's preferences for 6 universities: London School of Economics, HEC Paris, Università Commerciale Luigi Bocconi (Milano), Universität St. Gallen, ESADE (Barcelona), Handelshögskolan i Stockholm. To identify reasons for the preferences, several subject covariates (including foreign language competence, gender, etc.) have been assessed. Furthermore, several object covariates are attached to preference (and preference1998): the universities' field of specialization (economics, management science, finance) and location (Latin country, or other).

The correct data are available in the online complements to Dittrich et al. (1998). However, the accompanying analysis was based on an erroneous version of the data in which the choices for the last comparison pair (Barcelona : Stockholm) were accidentally reversed. See the corrigendum in Dittrich et al. (2001) for further details. The variable preference provides the correct data and can thus be used to replicate the analysis from the corrigendum (Dittrich et al. 2001). For convenience, the erroneous version is provided in preference1998 which can therefore be used to replicate the (incorrect) original analysis (Dittrich et al. 1998).

Source

The Royal Statistical Society Datasets Website.

References

Dittrich R, Hatzinger R, Katzenbeisser W (1998). Modelling the Effect of Subject-Specific Covariates in Paired Comparison Studies with an Application to University Rankings, Journal of the Royal Statistical Society C, 47, 511–525.

Dittrich R, Hatzinger R, Katzenbeisser W (2001). Corrigendum: Modelling the Effect of Subject-Specific Covariates in Paired Comparison Studies with an Application to University Rankings, Journal of the Royal Statistical Society C, 50, 247–249.

See Also

paircomp

Examples

data("CEMSChoice", package = "psychotree")
summary(CEMSChoice$preference)
covariates(CEMSChoice$preference)

[Package psychotree version 0.16-2 Index]