DIFSim {psychotree}	R Documentation

Artificial Data with Differential Item Functioning

Description

Artificial data simulated from a Rasch model and a partial credit model, respectively, where the items exhibit differential item functioning (DIF).

Usage

data(DIFSim)
data(DIFSimPC)

Format

Two data frames containing 200 and 500 observations, respectively, on 4 variables.

resp

an itemresp matrix with binary or polytomous results for 20 or 8 items, respectively.

age

age in years.

gender

factor indicating gender.

motivation

ordered factor indicating motivation level.

Details

The data are employed for illustrations in Strobl et al. (2015) and Komboz et al. (2018). See the manual pages for raschtree and pctree for fitting the tree models..

References

Komboz B, Zeileis A, Strobl C (2018). Tree-Based Global Model Tests for Polytomous Rasch Models. Educational and Psychological Measurement, 78(1), 128–166. doi: 10.1177/0013164416664394

Strobl C, Kopf J, Zeileis A (2015). Rasch Trees: A New Method for Detecting Differential Item Functioning in the Rasch Model. Psychometrika, 80(2), 289–316. doi: 10.1007/s11336-013-9388-3

See Also

raschtree, pctree

Examples

## data
data("DIFSim", package = "psychotree")
data("DIFSimPC", package = "psychotree")

## summary of covariates
summary(DIFSim[, -1])
summary(DIFSimPC[, -1])

## empirical frequencies of responses
plot(DIFSim$resp)
plot(DIFSimPC$resp)

## histogram of raw scores
hist(rowSums(DIFSim$resp), breaks = 0:20 - 0.5)
hist(rowSums(DIFSimPC$resp), breaks = 0:17 - 0.5)

[Package psychotree version 0.16-2 Index]