EuropeanValuesStudy {psychotree}	R Documentation

European Values Study

Description

A sample of the 1999 European Values Study (EVS) containing an assessment of materialism/postmaterialism in 3584 respondents from 32 countries.

Usage

data("EuropeanValuesStudy")

Format

A data frame containing 3584 observations on 10 variables.

country

Factor coding the country of a respondent.

gender

Factor coding gender.

birthyear

Numeric. Year of birth.

eduage

Numeric. Age when full time education was or will be completed.

marital

Factor. Current legal marital status.

employment

Ordered factor. Employment and number of working hours.

occupation

Factor. What is/was your main job?

income

Ordered factor. Income of household in ten categories from 10 percent lowest to 10 percent highest income category.

paircomp

Paired comparison of class paircomp. Five pairwise choices among four important political goals derived from a double-choice task (see Details).

country2

Factor. Country group according to postmaterialism (see Details).

Details

The data are part of a larger survey conducted in 1999 in 32 countries in Europe (see https://europeanvaluesstudy.eu/). Vermunt (2003) obtained a sample from 10 percent of the available cases per country, yielding 3584 valid cases.

The item in the 1999 European Values Study questionnaire aiming at recording materialism/postmaterialism reads as follows:

There is a lot of talk these days about what the aims of this country should be for the next ten years. On this card are listed some of the goals which different people would give top priority. If you had to choose, which of the things on this card would you say is most important? And which would be the next most important?

A Maintaining order in the nation
B Giving people more say in important government decisions
C Fighting rising prices
D Protecting freedom of speech

The double-choice task implies a partial ranking of the alternatives and (assuming transitivity) an incomplete set of paired comparisons for each respondent.

The country group according to postmaterialism was derived by Vermunt (2003) using a latent class model, and applied by Lee and Lee (2010) in a tree model.

Source

Latent GOLD Sample Data Sets Website.

References

Lee PH, Yu PLH (2010). Distance-Based Tree Models for Ranking Data. Computational Statistics and Data Analysis, 54, 1672–1682.

Vermunt JK (2003). Multilevel Latent Class Models. Sociological Methodology, 33, 213–239.

See Also

paircomp

Examples

## data
data("EuropeanValuesStudy", package = "psychotree")
summary(EuropeanValuesStudy$paircomp)

## Not run: 
## Bradley-Terry tree resulting in similar results compared to
## the (different) tree approach of Lee and Lee (2010)
evs <- na.omit(EuropeanValuesStudy)
bt <- bttree(paircomp ~ gender + eduage + birthyear + marital + employment + income + country2,
  data = evs, alpha = 0.01)
plot(bt, abbreviate = 2)

## End(Not run)

[Package psychotree version 0.16-2 Index]