gpcmtree {psychotree}	R Documentation

Generalized Partial Credit Model Trees

Description

Recursive partitioning (also known as trees) based on generalized partial credit models (GPCMs) for global testing of differential item functioning (DIF).

Usage

gpcmtree(formula, data, weights = NULL,
  grouppars = FALSE, vcov = TRUE, nullcats = "downcode",
  start = NULL, method = "BFGS", maxit = 500L,
  reltol = 1e-10, minsize = 500, ...)

## S3 method for class 'gpcmtree'
plot(x, type = c("regions", "profile"), terminal_panel = NULL,
  tp_args = list(...), tnex = 2L, drop_terminal = TRUE, ...)

Arguments

formula
A symbolic description of the model to be fit. This should be of type y ~ x1 + x2 where y should be an item response matrix and x1 and x2 are used as partitioning variables. Additionally, it is poosible to allow for impact of a group variable so that different ability distributions are estimated in each group. This can be specified by extending the previous formula by a group factor g as y ~ g | x1 + x2.

data
a data frame containing the variables in the model.

weights
an optional vector of weights (interpreted as case weights).

grouppars
logical. Should the estimated distributional group parameters of a multiple group model be included in the model parameters?

vcov
logical or character specifying the type of variance-covariance matrix (if any) computed for the final models (see gpcmodel).

nullcats
character string, specifying how items with null categories (i.e., categories not observed) should be treated. See gpcmodel, currently only "downcode" is available.

start
an optional vector or list of starting values (see gpcmodel).

method
control parameter for the optimizer employed by mirt for the EM algorithm (see gpcmodel).

maxit
control parameter for the optimizer employed by gpcmodel.

reltol
control parameter for the optimizer employed by gpcmodel.

minsize
integer specification of minimum number of observations in each node, which is passed to mob_control.

...
arguments passed to mob_control for gpcmtree, or to the underlying plot method, respectively.

x
an object of class gpcmtree.

type
character specifying the type of plot.

terminal_panel, tp_args, tnex, drop_terminal
arguments passed to mob.

Details

Generalized partial credit model (GPCM) trees are an application of model-based recursive partitioning (implemented in mob) to GPCM models (implemented in gpcmodel). For technical and algorithmic details, see the documentation of the two core functions linked above as well as vignette("mob", package = "partykit").

Various methods are provided for "gpcmtree" objects, most of them inherit their behavior from "modelparty" objects (e.g., print, summary). Additionally, dedicated extractor functions or provided for the different groups of model parameters in each node of the tree: itempar (item parameters), threshpar (threshold parameters), guesspar (guessing parameters), upperpar (upper asymptote parameters).

Value

An object of S3 class "gpcmtree" inheriting from class "modelparty".

See Also

mob, plmodel, rstree, pctree, raschtree, npltree

[Package psychotree version 0.16-2 Index]