node_btplot {psychotree}	R Documentation

Panel-Generating Function for Visualizing Bradley-Terry Tree Models

Description

Panel-generating function for visualizing the worth parameters from the nodes in Bradley-Terry tree models.

Usage

node_btplot(mobobj, id = TRUE,
  worth = TRUE, names = TRUE, abbreviate = TRUE, index = TRUE, ref = TRUE,
  col = "black", refcol = "lightgray", bg = "white", cex = 0.5, pch = 19,
  xscale = NULL, yscale = NULL, ylines = 1.5)

Arguments

mobobj
an object of class "mob" based on Bradley-Terry models fitted by btmodel.

id
logical. Should the node ID be displayed?

worth
logical. Should worth parameters (or their logs) be visualized?

names
logical. Should the names for the objects be displayed?

abbreviate
logical or numeric. Should object names be abbreviated? If numeric this controls the length of the abbreviation.

index
logical. Should different indexes for different stimuli be used?

ref
logical. Should a horizontal line for the reference level be drawn? Alternatively, ref can also be numeric or character to employ a reference level different from that stored in the model object.

col, cex, pch
graphical appearance of plotting symbols.

refcol
line color for reference line (if ref).

bg
color for background filling.

xscale, yscale
x and y axis limits.

ylines
numeric. Number of lines used for y-axis labels.

Details

The panel-generating function node_btplot is called by the plot method for "bttree" objects and does not have to be called by the user directly.

Value

A panel function which can be supplied to the plot method for "mob" objects.

See Also

bttree

[Package psychotree version 0.16-2 Index]