node_mptplot {psychotree}	R Documentation

Panel-Generating Function for Visualizing MPT Tree Models

Description

Panel-generating function for visualizing the model parameters from the nodes in MPT tree models.

Usage

node_mptplot(mobobj, id = TRUE,
  names = TRUE, abbreviate = TRUE, index = TRUE, ref = TRUE,
  col = "black", linecol = "lightgray", bg = "white", cex = 0.5, pch = 19,
  xscale = NULL, yscale = c(0, 1), ylines = 1.5)

Arguments

mobobj
an object of class "mob" based on MPT models fitted by mptmodel.

id
logical. Should the node ID be displayed?

names
logical or character. Should the names for the parameters be displayed? If character, this sets the names.

abbreviate
logical or numeric. Should parameter names be abbreviated? If numeric this controls the length of the abbreviation.

index
logical or character. Should different indexes for different parameters be used? If character, this controls the order of labels given in names.

ref
logical. Should a horizontal line for the reference level be drawn?

col, cex, pch
graphical appearance of plotting symbols.

linecol
line color for reference line (if ref).

bg
color for background filling.

xscale, yscale
x and y axis limits.

ylines
numeric. Number of lines used for y-axis labels.

Details

The panel-generating function node_mptplot is called by the plot method for "mpttree" objects and does not have to be called by the user directly.

Value

A panel function which can be supplied to the plot method for "mob" objects.

See Also

mpttree

[Package psychotree version 0.16-2 Index]