node_profileplot {psychotree}	R Documentation

Panel-Generating Function for Visualizing IRT Tree Models

Description

Panel-generating function for visualizing profiles (points and lines) of the parameters from the nodes in IRT tree models.

Usage

node_profileplot(
  mobobj,
  what = c("item", "coef", "threshold", "discrimination", "guessing", "upper"),
  parg = list(type = NULL, ref = NULL, alias = TRUE, logit = FALSE),
  id = TRUE,
  names = FALSE,
  abbreviate = TRUE,
  index = TRUE,
  ref = TRUE,
  col = "black",
  border = col,
  linecol = "black",
  refcol = "lightgray",
  bg = "white",
  cex = 0.5,
  pch = 21,
  xscale = NULL,
  yscale = NULL,
  ylines = 2,
  ...
)

Arguments

mobobj
an object of class "npltree" or class "mob" fitted by npltree

what
specifying the type of parameters to be plotted

parg
supplementary arguments for "what"

id
logical. Should the node ID be displayed?

names
logical or character. If TRUE, the names of the items are displayed on the x-axis. If FALSE, numbers of items are shown. Alternatively a character vector of the same length as the number of items can be supplied.

abbreviate
logical. Should item names be abbreviated? If numeric this controls the length of the abbreviation.

index
logical. Should different indexes for different items be used?

ref
logical. Should a horizontal line for the reference level be drawn?

col, border, pch, cex
graphical appearance of plotting symbols.

linecol, refcol
character, specifying the line color to use for the profile lines and reference line, respectively.

bg
color for background filling.

xscale, yscale
x and y axis limits.

ylines
numeric. Number of lines used for y-axis labels.

...
further arguments currently not used.

Details

The panel-generating function node_regionplot is called by the plot method of "gpcmtree" object by default and does not have to be called by the user directly. See regionplot for details and references of the drawn region plots and possible values and their meaning for the argument type (taken by node_regionplot).

Value

A panel function which can be supplied to the plot method for "npltree" objects or "mob" objects fitted by npltree or gpcmtree.

[Package psychotree version 0.16-2 Index]