node_regionplot {psychotree}	R Documentation

Panel-Generating Function for Visualizing IRT Tree Models

Description

Panel-generating function for visualizing the regions of expected item responses across abilities (via shaded rectangles) based on the parameters from the nodes in IRT tree models.

Usage

node_regionplot(
  mobobj,
  names = FALSE,
  abbreviate = TRUE,
  type = c("mode", "median", "mean"),
  ref = NULL,
  ylim = NULL,
  off = 0.1,
  col_fun = gray.colors,
  bg = "white",
  uo_show = TRUE,
  uo_col = "red",
  uo_lty = 2,
  uo_lwd = 1.25,
  ylines = 2
)

Arguments

mobobj
an object of class "npltree" or class "mob" fitted by npltree

names
logical or character. If TRUE, the names of the items are displayed on the x-axis. If FALSE, numbers of items are shown. Alternatively a character vector of the same length as the number of items can be supplied.

abbreviate
logical. Should item names be abbreviated? If numeric this controls the length of the abbreviation.

type
character, specifying which type of threshold parameters are to be used to mark the category regions per item in the plot (see regionplot for details).

ref
a vector of labels or position indices of item parameters which should be used as restriction/for normalization. If NULL (the default), all items are used (sum zero restriction). See threshpar for more details.

ylim
y axis limits

off
numeric, the distance (in scale units) between two item rectangles.

col_fun
function. Function to use for creating the color palettes for the rectangles. Per default gray.colors is used. Be aware that col_fun should accept as first argument an integer specifying the number of colors to create.

bg
color for background filling.

uo_show
logical. If set to TRUE (the default), disordered absolute item threshold parameters are indicated by a horizontal line (only if type is set to "mode").

uo_col
character, color of indication lines (if uo_show).

uo_lty
numeric. Line typ of indication lines (if uo_show).

uo_lwd
numeric. Line width of indication lines (if uo_show).

ylines
numeric. Number of lines used for y-axis labels.

Value

A panel function which can be supplied to the plot method for "npltree" objects or "mob" objects fitted by npltree.

[Package psychotree version 0.16-2 Index]